requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.52.1
kiwisolver==1.4.5
matplotlib==3.9.0
numpy==1.26.4
packaging==24.0
pillow==10.3.0
pyparsing==3.1.2
pyserial==3.5
python-dateutil==2.9.0.post0
six==1.16.0
inference_sdk==0.11.2
roboticstoolbox_python==1.1.1
# qpsolvers==4.3.2
